week3/pyproject.toml
[project]
name = "week3-multi-agent"
version = "0.1.0"
description = "Multi-agent hat store customer support with ADK, MCP, and A2A"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "httpx>=0.28.0",
    "litellm>=1.83.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
addopts = "-m 'not eval'"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_api_key: marks tests needing ANTHROPIC_API_KEY and Supabase",
    "eval: LLM-as-judge evals, excluded from default runs",
]
